refactor(people): split linker helpers
Move name policy, fuzzy matching, OCD IDs, PDF lookup, person selection, and mutation helpers behind focused modules.

Keep pipeline.utils and pipeline.person_linker imports stable while preserving matching thresholds, cache behavior, and reindex side effects.
